Imported newest cram code from Staden Package. This primarily consists of renaming sam_header_* to sam_hdr_* allowing libstaden-read and libbam to be linked into the same application.
In light of this we don't need as much #define hackery to avoid
clashes here either.

Also enabled -DHAVE_CURL in the Makefile and updated link lines in
order to avoid the clumsy use of fork/execlp wget (and to avoid a
whinge about tempnam, but that's a different issue).