Rename het_phred to het_logodd.
It's confusing to call it a phred score when it isn't. It's log(p/(1-p)).

Also improved the accuracy of the phred calculation.  The maths it
hand tended to the correct score for high values, but not for low
ones.  We know when it'll be wrong so we can switch when necessary to
the more complex expression.