Expression strings are now kstring.
This permits us to have constant values in queries as the string isn't
modified anymore, and gives room for future string manipulation such
as removal of backslashes and escaping of quotes.

We also now have an explicit "is_str" flag.
This gives us the advantage of better distinguishing between an empty
string and a non-existant string.